Extract URL arbitration as a controller concern.
